+++ packages/did-core-test-server/src/services/scenarios/did-parameters/did-parameters-positive-tests.js
@@ -1,4 +1,8 @@
-const { isAsciiString, getQueryParamValueFromDidUri } = require('../../utils');
+const {
+  isAsciiString,
+  getQueryParamValueFromDidUri,
+  generateScenarioResults,
+} = require('../../utils');
 
 const assertions = {
   '[hl] The associated value MUST be an ASCII string.': (scenario) => {
@@ -39,24 +43,4 @@ const assertions = {
     return value === decoded && decoded === '/path/to/resource';
   },
 };
-
-module.exports = (scenario) => {
-  let assertion_results = {};
-
-  Object.keys(assertions).forEach((assertion) => {
-    assertion_results = {
-      ...assertion_results,
-      [assertion]: assertions[assertion](scenario),
-    };
-  });
-
-  const test_result = Object.values(assertion_results).every((result) => {
-    return result === true;
-  });
-
-  return {
-    scenario: scenario.name,
-    test: test_result ? 'PASS' : 'FAIL',
-    assertion_results,
-  };
-};
+module.exports = generateScenarioResults(assertions);
